requirements.txt
python-dotenv
ultralytics
fastapi[all]
